requirements.txt
torch
torchvision
segmentation-models-pytorch
pandas
numpy
Pillow
opencv-python
matplotlib
tqdm
scikit-learn
